Removing the validation check on floating type inputs that tries to validate that inputs that look like integers must be exactly representable as a FP number.
For instance, if the user sends, "1435774380" as an input that is a DT_FLOAT type, the previous behavior would return an error because 1435774380 "looks like" an int and cannot be exactly represented as an integer as a DT_FLOAT.

But this behavior does not make sense. For instance, "1435774380.0" is exactly equivalent to "1435774380" and yet the former is accepted while the latter returns an error. Since the user clearly set the data type to DT_FLOAT, this is surprising and even undesirable behavior (see below). Imagine if `atof` in C correctly parsed the former but rejected the latter. That would be confusing to users, and sometimes make the function difficult or impossible to use.

One could argue that the user can "opt-in" to the desired behavior by choosing whether or not to use ".0" at the end of the inputs. While this is true in theory, it is not true in practice since underlying libraries that generate JSON may convert something like `double d = 1435774380.0;` to "1435774380". In fact, this is _exactly_ what happens in Google's own Protobuf -> JSON conversion with the user having absolutely no control over the behavior (FWIW, the library uses `snprintf` with the format string "%.*g"). So users of Google's own libraries cannot generate JSON that is usable by tf-serving.

PiperOrigin-RevId: 364832510